requirements.txt
pytorch==0.4.1
librosa==0.7.1
pydub
pyaudio==0.2.11
sklearn==0.21.3
numpy==1.17.3
pandas==0.24.2
matplotlib==3.1.0
seaborn==0.9.0